refact: Extract a base component for model lists
